Got user to log in.
