+++ app/build.gradle
@@ -43,4 +43,12 @@ dependencies {
 
     // Timer
     implementation 'com.github.anugotta:FlipTimerView:v1.5'
+
+
+    //Internet
+    implementation 'com.squareup.retrofit2:retrofit:2.3.0'
+    implementation 'com.squareup.retrofit2:converter-gson:2.3.0'
+    implementation 'com.google.code.gson:gson:2.8.6'
+    implementation 'com.squareup.okhttp3:logging-interceptor:3.9.0'
+
 }
