+++ app/src/main/AndroidManifest.xml
@@ -2,9 +2,14 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     package="io.brickhack.mobile">
+
     <uses-permission android:name="android.permission.NFC" />
     <uses-permission android:name="android.permission.VIBRATE" />
-    <uses-feature android:name="android.hardware.nfc" android:required="true"/>
+
+    <uses-feature
+        android:name="android.hardware.nfc"
+        android:required="true" />
+
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
@@ -14,15 +19,20 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name="io.brickhack.mobile.unused.Authentication" android:label="@string/app_name">
+        <activity
+            android:name=".Activities.LoginActivity"
+            android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
             <intent-filter>
-                <action android:name="io.brickhack.mobile.appauth.HANDLE_AUTHORIZATION_RESPONSE"/>
-                <category android:name="android.intent.category.DEFAULT"/>
-                <data android:scheme="brickhack"/>
+                <action android:name="io.brickhack.mobile.appauth.HANDLE_AUTHORIZATION_RESPONSE" />
+
+                <category android:name="android.intent.category.DEFAULT" />
+
+                <data android:scheme="brickhack" />
             </intent-filter>
         </activity>
         <activity android:name="net.openid.appauth.RedirectUriReceiverActivity">
@@ -33,9 +43,11 @@
                 <data android:scheme="brickhack"/>
             </intent-filter>
         </activity>
-        <activity android:name=".unused.DashBoard"/>
-        <activity android:name=".unused.Wristband"/>
-        <activity android:name=".unused.History"/>
-        <activity android:name=".unused.ManageAttendees"/>
+        <activity android:name=".MainActivity"/>
+        <!--<activity android:name=".unused.DashBoard" />-->
+        <!--<activity android:name=".unused.Wristband" />-->
+        <!--<activity android:name=".unused.History" />-->
+        <!--<activity android:name=".unused.ManageAttendees" />-->
     </application>
+
 </manifest>
