+++ app/src/main/java/io/brickhack/mobile/Settings.java
@@ -1,11 +1,10 @@
-package io.brickhack.mobile.unused;
+package io.brickhack.mobile;
 
 import android.net.Uri;
 
-public class BrickHackSettings{
-
-    private final String clientID = "b181105e8a2bcf3552eea153be952981ff6cc5d9746a5161e526302e97780cd0";
-    private final String clientSCRT = "912ae062cbd507abb38191f23d4d8703d151fec09653001de4deedafa6e766b0";
+public class Settings {
+    private final String clientID = "b0a484dfaf474fdfd43ad7867d3c70fe8d76195ee565f36a29677fdbd8a168d3";
+    private final String clientSCRT = "1837394c4b13cf3954488a31afc2ec9482a33b2b7e68568280ae74edb954d756";
     private final String redirectURI = "brickhack://oauth/callback";
     public static final String LOG_TAG = "BrickHack";
