--- app/src/main/java/io/brickhack/mobile/unused/History.java
@@ -1,19 +1,0 @@
-package io.brickhack.mobile.unused;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
-
-import io.brickhack.mobile.R;
-
-public class History extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_history);
-        Toolbar toolbar = findViewById(R.id.toolbar);
-
-        setSupportActionBar(toolbar);
-    }
-}
