--- app/src/main/java/io/brickhack/mobile/unused/backendTag.java
@@ -1,15 +1,0 @@
-package io.brickhack.mobile.unused;
-
-import android.support.annotation.NonNull;
-
-public class backendTag {
-
-    Integer id;
-    String name;
-
-    @NonNull
-    @Override
-    public String toString() {
-        return name;
-    }
-}
