--- app/src/main/java/io/brickhack/mobile/unused/postTagFormat.java
@@ -1,13 +1,0 @@
-package io.brickhack.mobile.unused;
-
-public class postTagFormat {
-
-    private String band_id;
-    private int trackable_tag_id;
-
-    public postTagFormat(String band_id, int trackable_tag_id){
-        this.band_id = band_id;
-        this.trackable_tag_id = trackable_tag_id;
-    }
-
-}
