+++ app/src/main/res/layout/activity_main.xml
@@ -25,103 +25,23 @@
         app:layout_constraintVertical_bias="0.009"
         android:src="@drawable/hack_logo"/>
 
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginTop="32dp"
-        android:gravity="center"
-        android:text="@string/brickhack_date"
-        android:textSize="20sp"
-        android:textColor="@android:color/white"
-        app:layout_constraintTop_toBottomOf="@+id/id_logo"
-        tools:layout_editor_absoluteX="0dp" />
 
     <TextView
-        android:id="@+id/textView2"
-        android:layout_width="match_parent"
+        android:id="@+id/username"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:gravity="center"
-        android:text="@string/brickhack_location"
-        android:textSize="20sp"
-        android:textColor="@android:color/white"
-        app:layout_constraintTop_toBottomOf="@+id/textView"
-        tools:layout_editor_absoluteX="8dp" />
-
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
-        android:text="@string/sign_in"
-        android:textSize="25sp"
-        android:gravity="center"
-        android:textColor="@android:color/white"
-        app:layout_constraintBottom_toTopOf="@+id/textInputLayout"
-        app:layout_constraintTop_toBottomOf="@+id/textView2"
-        tools:layout_editor_absoluteX="117dp" />
-
-
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/textInputLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="6dp"
-        android:layout_marginBottom="8dp"
-        app:layout_constraintBottom_toTopOf="@+id/id_logo_button"
-        app:layout_constraintTop_toBottomOf="@+id/textView2"
-        app:layout_constraintVertical_bias="0.53">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/brickhack_username"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/username"
-            android:inputType="text"
-            android:layout_marginStart="20dp"
-            android:layout_marginEnd="20dp"
-            android:textColorHint="@android:color/white"/>
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <com.google.android.material.textfield.TextInputLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="6dp"
-        android:layout_marginBottom="8dp"
-        app:layout_constraintBottom_toTopOf="@+id/id_logo_button"
-        app:layout_constraintTop_toBottomOf="@+id/textView2"
-        app:layout_constraintVertical_bias="0.84000003">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/brickhack_password"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/password"
-            android:textColorHint="@android:color/white"
-            android:layout_marginStart="20dp"
-            android:layout_marginEnd="20dp"
-            android:inputType="textPassword"/>
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <Button
-        android:id="@+id/id_logo_button"
-        android:layout_width="300dp"
-        android:layout_height="47dp"
         android:layout_marginStart="8dp"
         android:layout_marginTop="8dp"
         android:layout_marginEnd="8dp"
         android:layout_marginBottom="8dp"
-        android:text="@string/login"
-        android:textAllCaps="true"
-        android:background="@android:color/holo_orange_dark"
-        android:textColor="@android:color/white"
-        android:textSize="20sp"
-        android:visibility="visible"
+        android:text="Main Activity"
+        android:textColor="@color/white"
+        android:textSize="30sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/id_logo"
-        app:layout_constraintVertical_bias="0.815" />
+        app:layout_constraintVertical_bias="0.405" />
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
