+++ app/src/main/res/values/colors.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<resources>
+    <resources>
     <color name="title_bar_color">#47525f</color>
     <color name="colorPrimaryDark">#303F9F</color>
     <color name="colorAccent">#FF4081</color>
@@ -8,3 +8,4 @@
     <color name="white">#ffffff</color>
     <color name="scan_bar">#759e9e9e</color>
 </resources>
+
